+++ crates/biome_js_analyze/tests/specs/nursery/noInvalidUseBeforeDeclaration/valid.js
@@ -10,3 +10,19 @@ export { X }; const X = 1;
 let a; console.log(a);
 
 function h() { Y; }; const Y = 0;
+
+const aFunction = (a = '111', b = a) => {
+    console.info(a,b);
+}
+
+const aFunction = (b = a) => {
+    console.info(a,b);
+}
+
+const aFunction = ({a = '111', b = a}) => {
+    console.info(a,b);
+}
+
+const aFunction = ({b = a}) => {
+    console.info(a,b);
+}